ID-6491 [TEMP] Fix rtbrick_confd_sock_access for TACACS users
In BUG 6491 the issue is that a TACACS user logging in does not have
any of it's supplementary groups from /etc/group set. One of these
supplementary groups might be rtbrick_confd_sock_access which is needed
for CLI2 usage. So usually the broken behavior that we see is that the
TACACS user can successfully login but cannot use CLI2. However this is
just one possible behavior, based on the contents of /etc/group there
might be other behaviors broken due to 6491.

With this change we add a fix which is specific for
rtbrick_confd_sock_access via the pam_group module and
/etc/security/group.conf . This change basically hard-codes that any
member of supervisor, operator and reader groups have CLI2 access.

In a default system configuration this is the desired behavior however
with more advanced configurations this might not be what the customer
wants.

This BREAKS the "no-cli-access" configuration knob.